Ajustando formatação de saídas dos serviços de atividades complementares e faltas
